Added information to README, merge of *API into DiscovererAPI, changed FetchIfAPI from POST to GET
