+++ discoverer/README.md
@@ -1,7 +1,8 @@
 SeaClouds Discoverer
 ====================
 
-The SeaClouds Discoverer module provides information about cloud offering used by SeaCloudsPlatform.
+The SeaClouds Discoverer module is able to provide information about cloud offerings through its RESTful API.
+Those information are a TOSCA YAML representation of the offerings retrieved by the crawlers from [CloudHarmony](https://cloudharmony.com) and [PaaSify](http://www.paasify.it).
 
 # Installing and running SeaClouds Discoverer
 The easiest way to install and run this component is explained in SeaClouds [README.md](../README.md).
@@ -9,8 +10,8 @@ The easiest way to install and run this component is explained in SeaClouds [REA
 ## Requirements
 Running SeaClouds Discoverer only requires Java 7 (or greater) installed on the target machine. Among with the software requirements, it also needs:
 
-- A discovererconf.yml configuration file.
-- A running MongoDB service
+- A discovererconf.yml configuration file
+- A running MongoDB service, used as a permanent layer to store information about cloud offerings used by SeaCloudsPlatform
 
 ### Config.yml configuration file
 SeaClouds Discoverer is packaged as a Dropwizard application. It requires a configuration file to run. SeaClouds already has an example [discovererconf.yml](./discovererconf.yml)
