+++ discoverer/src/main/java/eu/seaclouds/platform/discoverer/service/DiscovererApplication.java
@@ -20,11 +20,6 @@ public class DiscovererApplication extends Application<DiscovererConfiguration> 
 
         Discoverer discoverer = new Discoverer(mongoClient, configuration.getActiveCrawlers());
 
-        environment.jersey().register(new FetchAPI(discoverer));          /* /fetchOffer */
-        environment.jersey().register(new FetchAllAPI(discoverer));       /* /fetch_all */
-        environment.jersey().register(new DeleteAPI(discoverer));         /* /delete */
-        environment.jersey().register(new StatisticsAPI(discoverer));     /* /statistics */
-        environment.jersey().register(new RefreshAPI(discoverer));        /* /refresh */
-        environment.jersey().register(new FetchIfAPI(discoverer));
+        environment.jersey().register(new DiscovererAPI(discoverer));
     }
 }
